+++ test/e2e/metamask-ui.spec.js
@@ -10,7 +10,6 @@ const {
   closeAllWindowHandlesExcept,
   findElement,
   findElements,
-  loadExtension,
   openNewPage,
   switchToWindowWithTitle,
   verboseReportOnFailure,
@@ -18,9 +17,9 @@ const {
   setupFetchMocking,
   prepareExtensionForTesting,
 } = require('./helpers')
+const enLocaleMessages = require('../../app/_locales/en/messages.json')
 
 describe('MetaMask', function () {
-  let extensionId
   let driver
   let tokenAddress
 
@@ -35,7 +34,6 @@ describe('MetaMask', function () {
   before(async function () {
     const result = await prepareExtensionForTesting()
     driver = result.driver
-    extensionId = result.extensionId
     await setupFetchMocking(driver)
   })
 
@@ -60,7 +58,7 @@ describe('MetaMask', function () {
   describe('Going through the first time flow', () => {
     it('clicks the continue button on the welcome screen', async () => {
       await findElement(driver, By.css('.welcome-page__header'))
-      const welcomeScreenBtn = await findElement(driver, By.css('.first-time-flow__button'))
+      const welcomeScreenBtn = await findElement(driver, By.xpath(`//button[contains(text(), '${enLocaleMessages.getStarted.message}')]`))
       welcomeScreenBtn.click()
       await delay(largeDelayMs)
     })
@@ -105,7 +103,7 @@ describe('MetaMask', function () {
       assert.equal(seedPhrase.split(' ').length, 12)
       await delay(regularDelayMs)
 
-      const nextScreen = (await findElements(driver, By.css('button.first-time-flow__button')))[1]
+      const nextScreen = await findElement(driver, By.xpath(`//button[contains(text(), '${enLocaleMessages.next.message}')]`))
       await nextScreen.click()
       await delay(regularDelayMs)
     })
@@ -118,37 +116,12 @@ describe('MetaMask', function () {
       await delay(tinyDelayMs)
     }
 
-    async function retypeSeedPhrase (words, wasReloaded, count = 0) {
-      try {
-        if (wasReloaded) {
-          const byRevealButton = By.css('.reveal-seed-phrase__secret-blocker .reveal-seed-phrase__reveal-button')
-          await driver.wait(until.elementLocated(byRevealButton, 10000))
-          const revealSeedPhraseButton = await findElement(driver, byRevealButton, 10000)
-          await revealSeedPhraseButton.click()
-          await delay(regularDelayMs)
-
-          const nextScreen = await findElement(driver, By.css('button.first-time-flow__button'))
-          await nextScreen.click()
-          await delay(regularDelayMs)
-        }
-
-        for (let i = 0; i < 12; i++) {
-          await clickWordAndWait(words[i])
-        }
-      } catch (e) {
-        if (count > 2) {
-          throw e
-        } else {
-          await loadExtension(driver, extensionId)
-          await retypeSeedPhrase(words, true, count + 1)
-        }
-      }
-    }
-
     it('can retype the seed phrase', async () => {
       const words = seedPhrase.split(' ')
 
-      await retypeSeedPhrase(words)
+      for (const word of words) {
+        await clickWordAndWait(word)
+      }
 
       const confirm = await findElement(driver, By.xpath(`//button[contains(text(), 'Confirm')]`))
       await confirm.click()
@@ -157,7 +130,7 @@ describe('MetaMask', function () {
 
     it('clicks through the success screen', async () => {
       await findElement(driver, By.xpath(`//div[contains(text(), 'Congratulations')]`))
-      const doneButton = await findElement(driver, By.css('button.first-time-flow__button'))
+      const doneButton = await findElement(driver, By.xpath(`//button[contains(text(), '${enLocaleMessages.endOfFlowMessage10.message}')]`))
       await doneButton.click()
       await delay(regularDelayMs)
     })
@@ -249,7 +222,7 @@ describe('MetaMask', function () {
 
       await passwordInputs[0].sendKeys('correct horse battery staple')
       await passwordInputs[1].sendKeys('correct horse battery staple')
-      await driver.findElement(By.css('.first-time-flow__button')).click()
+      await driver.findElement(By.xpath(`//button[contains(text(), '${enLocaleMessages.restore.message}')]`)).click()
       await delay(regularDelayMs)
     })
 
@@ -475,13 +448,13 @@ describe('MetaMask', function () {
       await approveButton.click()
 
       await driver.switchTo().window(dapp)
-      await delay(regularDelayMs)
+      await delay(2000)
     })
 
     it('initiates a send from the dapp', async () => {
       const send3eth = await findElement(driver, By.xpath(`//button[contains(text(), 'Send')]`), 10000)
       await send3eth.click()
-      await delay(5000)
+      await delay(2000)
 
       windowHandles = await driver.getAllWindowHandles()
       await switchToWindowWithTitle(driver, 'MetaMask Notification', windowHandles)
@@ -496,17 +469,15 @@ describe('MetaMask', function () {
 
       await gasPriceInput.sendKeys(Key.BACK_SPACE)
       await delay(50)
-      await gasPriceInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
       await gasPriceInput.sendKeys('10')
       await delay(50)
       await delay(tinyDelayMs)
       await delay(50)
       await gasLimitInput.sendKeys(Key.chord(Key.CONTROL, 'a'))
       await delay(50)
-
       await gasLimitInput.sendKeys('25000')
-      await delay(largeDelayMs * 2)
+
+      await delay(1000)
 
       const confirmButton = await findElement(driver, By.xpath(`//button[contains(text(), 'Confirm')]`), 10000)
       await confirmButton.click()
@@ -792,14 +763,12 @@ describe('MetaMask', function () {
       await modalTabs[1].click()
       await delay(regularDelayMs)
 
-      const [gasPriceInput, gasLimitInput] = await findElements(driver, By.css('.advanced-tab__gas-edit-row__input'))
+      const [gasPriceInput, gasLimitInput] = await findElements(driver, By.css('.advanced-gas-inputs__gas-edit-row__input'))
       const gasLimitValue = await gasLimitInput.getAttribute('value')
       assert(Number(gasLimitValue) < 100000, 'Gas Limit too high')
       await gasPriceInput.sendKeys(Key.chord(Key.CONTROL, 'a'))
       await delay(50)
 
-      await gasPriceInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
       await gasPriceInput.sendKeys(Key.BACK_SPACE)
       await delay(50)
       await gasPriceInput.sendKeys('10')
@@ -808,18 +777,9 @@ describe('MetaMask', function () {
       await delay(50)
       await gasLimitInput.sendKeys(Key.BACK_SPACE)
       await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
       await gasLimitInput.sendKeys('60001')
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.chord(Key.CONTROL, 'e'))
-      await delay(50)
+
+      await delay(1000)
 
       const save = await findElement(driver, By.xpath(`//button[contains(text(), 'Save')]`))
       await save.click()
@@ -914,7 +874,7 @@ describe('MetaMask', function () {
       await advancedTabButton.click()
       await delay(tinyDelayMs)
 
-      const [gasPriceInput, gasLimitInput] = await findElements(driver, By.css('.advanced-tab__gas-edit-row__input'))
+      const [gasPriceInput, gasLimitInput] = await findElements(driver, By.css('.advanced-gas-inputs__gas-edit-row__input'))
       assert(gasPriceInput.getAttribute('value'), 20)
       assert(gasLimitInput.getAttribute('value'), 4700000)
 
@@ -1103,12 +1063,10 @@ describe('MetaMask', function () {
       await modalTabs[1].click()
       await delay(regularDelayMs)
 
-      const [gasPriceInput, gasLimitInput] = await findElements(driver, By.css('.advanced-tab__gas-edit-row__input'))
+      const [gasPriceInput, gasLimitInput] = await findElements(driver, By.css('.advanced-gas-inputs__gas-edit-row__input'))
       await gasPriceInput.sendKeys(Key.chord(Key.CONTROL, 'a'))
       await delay(50)
 
-      await gasPriceInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
       await gasPriceInput.sendKeys(Key.BACK_SPACE)
       await delay(50)
       await gasPriceInput.sendKeys('10')
@@ -1117,18 +1075,9 @@ describe('MetaMask', function () {
       await delay(50)
       await gasLimitInput.sendKeys(Key.BACK_SPACE)
       await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
       await gasLimitInput.sendKeys('60000')
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.chord(Key.CONTROL, 'e'))
-      await delay(50)
+
+      await delay(1000)
 
       const save = await findElement(driver, By.css('.page-container__footer-button'))
       await save.click()
@@ -1246,12 +1195,10 @@ describe('MetaMask', function () {
       await modalTabs[1].click()
       await delay(regularDelayMs)
 
-      const [gasPriceInput, gasLimitInput] = await findElements(driver, By.css('.advanced-tab__gas-edit-row__input'))
+      const [gasPriceInput, gasLimitInput] = await findElements(driver, By.css('.advanced-gas-inputs__gas-edit-row__input'))
       await gasPriceInput.sendKeys(Key.chord(Key.CONTROL, 'a'))
       await delay(50)
 
-      await gasPriceInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
       await gasPriceInput.sendKeys(Key.BACK_SPACE)
       await delay(50)
       await gasPriceInput.sendKeys('10')
@@ -1260,18 +1207,9 @@ describe('MetaMask', function () {
       await delay(50)
       await gasLimitInput.sendKeys(Key.BACK_SPACE)
       await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.BACK_SPACE)
-      await delay(50)
       await gasLimitInput.sendKeys('60001')
-      await delay(50)
-      await gasLimitInput.sendKeys(Key.chord(Key.CONTROL, 'e'))
-      await delay(50)
+
+      await delay(1000)
 
       const save = await findElement(driver, By.css('.page-container__footer-button'))
       await save.click()