+++ ui/app/components/app/gas-customization/advanced-gas-inputs/advanced-gas-inputs.component.js
@@ -3,21 +3,10 @@ import PropTypes from 'prop-types'
 import classnames from 'classnames'
 import debounce from 'lodash.debounce'
 
-export default class AdvancedTabContent extends Component {
+export default class AdvancedGasInputs extends Component {
   static contextTypes = {
     t: PropTypes.func,
   }
-
-  constructor (props) {
-    super(props)
-    this.state = {
-      gasPrice: this.props.customGasPrice,
-      gasLimit: this.props.customGasLimit,
-    }
-    this.changeGasPrice = debounce(this.changeGasPrice, 500)
-    this.changeGasLimit = debounce(this.changeGasLimit, 500)
-  }
-
 
   static propTypes = {
     updateCustomGasPrice: PropTypes.func,
@@ -29,6 +18,16 @@ export default class AdvancedTabContent extends Component {
     isSpeedUp: PropTypes.bool,
     showGasPriceInfoModal: PropTypes.func,
     showGasLimitInfoModal: PropTypes.func,
+  }
+
+  constructor (props) {
+    super(props)
+    this.state = {
+      gasPrice: this.props.customGasPrice,
+      gasLimit: this.props.customGasLimit,
+    }
+    this.changeGasPrice = debounce(this.changeGasPrice, 500)
+    this.changeGasLimit = debounce(this.changeGasLimit, 500)
   }
 
   componentDidUpdate (prevProps) {
@@ -50,12 +49,7 @@ export default class AdvancedTabContent extends Component {
   }
 
   changeGasLimit = (e) => {
-    if (e.target.value < 21000) {
-      this.setState({ gasLimit: 21000 })
-      this.props.updateCustomGasLimit(21000)
-    } else {
-      this.props.updateCustomGasLimit(Number(e.target.value))
-    }
+    this.props.updateCustomGasLimit(Number(e.target.value))
   }
 
   onChangeGasPrice = (e) => {
@@ -67,89 +61,83 @@ export default class AdvancedTabContent extends Component {
     this.props.updateCustomGasPrice(Number(e.target.value))
   }
 
-  gasInputError ({ labelKey, insufficientBalance, customPriceIsSafe, isSpeedUp, value }) {
+  gasPriceError ({ insufficientBalance, customPriceIsSafe, isSpeedUp, gasPrice }) {
     const { t } = this.context
-    let errorText
-    let errorType
-    let isInError = true
-
 
     if (insufficientBalance) {
-      errorText = t('insufficientBalance')
-      errorType = 'error'
-    } else if (labelKey === 'gasPrice' && isSpeedUp && value === 0) {
-      errorText = t('zeroGasPriceOnSpeedUpError')
-      errorType = 'error'
-    } else if (labelKey === 'gasPrice' && !customPriceIsSafe) {
-      errorText = t('gasPriceExtremelyLow')
-      errorType = 'warning'
-    } else {
-      isInError = false
+      return {
+        errorText: t('insufficientBalance'),
+        errorType: 'error',
+      }
+    } else if (isSpeedUp && gasPrice === 0) {
+      return {
+        errorText: t('zeroGasPriceOnSpeedUpError'),
+        errorType: 'error',
+      }
+    } else if (!customPriceIsSafe) {
+      return {
+        errorText: t('gasPriceExtremelyLow'),
+        errorType: 'warning',
+      }
     }
 
-    return {
-      isInError,
-      errorText,
-      errorType,
+    return {}
+  }
+
+  gasLimitError ({ insufficientBalance, gasLimit }) {
+    const { t } = this.context
+
+    if (insufficientBalance) {
+      return {
+        errorText: t('insufficientBalance'),
+        errorType: 'error',
+      }
+    } else if (gasLimit < 21000) {
+      return {
+        errorText: t('gasLimitTooLow'),
+        errorType: 'error',
+      }
     }
+
+    return {}
   }
 
-  gasInput ({ labelKey, value, onChange, insufficientBalance, customPriceIsSafe, isSpeedUp }) {
-    const {
-      isInError,
-      errorText,
-      errorType,
-    } = this.gasInputError({ labelKey, insufficientBalance, customPriceIsSafe, isSpeedUp, value })
-
-    return (
-      <div className="advanced-gas-inputs__gas-edit-row__input-wrapper">
-        <input
-          className={classnames('advanced-gas-inputs__gas-edit-row__input', {
-            'advanced-gas-inputs__gas-edit-row__input--error': isInError && errorType === 'error',
-            'advanced-gas-inputs__gas-edit-row__input--warning': isInError && errorType === 'warning',
-          })}
-          type="number"
-          value={value}
-          onChange={onChange}
-        />
-        <div className={classnames('advanced-gas-inputs__gas-edit-row__input-arrows', {
-          'advanced-gas-inputs__gas-edit-row__input--error': isInError && errorType === 'error',
-          'advanced-gas-inputs__gas-edit-row__input--warning': isInError && errorType === 'warning',
-        })}>
-          <div
-            className="advanced-gas-inputs__gas-edit-row__input-arrows__i-wrap"
-            onClick={() => onChange({ target: { value: value + 1 } })}
-          >
-            <i className="fa fa-sm fa-angle-up" />
-          </div>
-          <div
-            className="advanced-gas-inputs__gas-edit-row__input-arrows__i-wrap"
-            onClick={() => onChange({ target: { value: Math.max(value - 1, 0) } })}
-          >
-            <i className="fa fa-sm fa-angle-down" />
-          </div>
-        </div>
-        { isInError
-          ? <div className={`advanced-gas-inputs__gas-edit-row__${errorType}-text`}>
-            { errorText }
-          </div>
-          : null }
-      </div>
-    )
-  }
-
-  infoButton (onClick) {
-    return <i className="fa fa-info-circle" onClick={onClick} />
-  }
-
-  renderGasEditRow (gasInputArgs) {
+  renderGasInput ({ value, onChange, errorComponent, errorType, infoOnClick, label }) {
     return (
       <div className="advanced-gas-inputs__gas-edit-row">
         <div className="advanced-gas-inputs__gas-edit-row__label">
-          { this.context.t(gasInputArgs.labelKey) }
-          { this.infoButton(() => gasInputArgs.infoOnClick()) }
+          { label }
+          <i className="fa fa-info-circle" onClick={infoOnClick} />
         </div>
-        { this.gasInput(gasInputArgs) }
+        <div className="advanced-gas-inputs__gas-edit-row__input-wrapper">
+          <input
+            className={classnames('advanced-gas-inputs__gas-edit-row__input', {
+              'advanced-gas-inputs__gas-edit-row__input--error': errorType === 'error',
+              'advanced-gas-inputs__gas-edit-row__input--warning': errorType === 'warning',
+            })}
+            type="number"
+            value={value}
+            onChange={onChange}
+          />
+          <div className={classnames('advanced-gas-inputs__gas-edit-row__input-arrows', {
+            'advanced-gas-inputs__gas-edit-row__input--error': errorType === 'error',
+            'advanced-gas-inputs__gas-edit-row__input--warning': errorType === 'warning',
+          })}>
+            <div
+              className="advanced-gas-inputs__gas-edit-row__input-arrows__i-wrap"
+              onClick={() => onChange({ target: { value: value + 1 } })}
+            >
+              <i className="fa fa-sm fa-angle-up" />
+            </div>
+            <div
+              className="advanced-gas-inputs__gas-edit-row__input-arrows__i-wrap"
+              onClick={() => onChange({ target: { value: Math.max(value - 1, 0) } })}
+            >
+              <i className="fa fa-sm fa-angle-down" />
+            </div>
+          </div>
+          { errorComponent }
+        </div>
       </div>
     )
   }
@@ -162,25 +150,47 @@ export default class AdvancedTabContent extends Component {
       showGasPriceInfoModal,
       showGasLimitInfoModal,
     } = this.props
+    const {
+      gasPrice,
+      gasLimit,
+    } = this.state
+
+    const {
+      errorText: gasPriceErrorText,
+      errorType: gasPriceErrorType,
+    } = this.gasPriceError({ insufficientBalance, customPriceIsSafe, isSpeedUp, gasPrice })
+    const gasPriceErrorComponent = gasPriceErrorType ?
+      <div className={`advanced-gas-inputs__gas-edit-row__${gasPriceErrorType}-text`}>
+        { gasPriceErrorText }
+      </div> :
+      null
+
+    const {
+      errorText: gasLimitErrorText,
+      errorType: gasLimitErrorType,
+    } = this.gasLimitError({ insufficientBalance, gasLimit })
+    const gasLimitErrorComponent = gasLimitErrorType ?
+      <div className={`advanced-gas-inputs__gas-edit-row__${gasLimitErrorType}-text`}>
+        { gasLimitErrorText }
+      </div> :
+      null
 
     return (
       <div className="advanced-gas-inputs__gas-edit-rows">
-        { this.renderGasEditRow({
-          labelKey: 'gasPrice',
+        { this.renderGasInput({
+          label: this.context.t('gasPrice'),
           value: this.state.gasPrice,
           onChange: this.onChangeGasPrice,
-          insufficientBalance,
-          customPriceIsSafe,
-          showGWEI: true,
-          isSpeedUp,
+          errorComponent: gasPriceErrorComponent,
+          errorType: gasPriceErrorType,
           infoOnClick: showGasPriceInfoModal,
         }) }
-        { this.renderGasEditRow({
-          labelKey: 'gasLimit',
+        { this.renderGasInput({
+          label: this.context.t('gasLimit'),
           value: this.state.gasLimit,
           onChange: this.onChangeGasLimit,
-          insufficientBalance,
-          customPriceIsSafe,
+          errorComponent: gasLimitErrorComponent,
+          errorType: gasLimitErrorType,
           infoOnClick: showGasLimitInfoModal,
         }) }
       </div>