+++ ui/app/components/app/gas-customization/gas-price-button-group/tests/gas-price-button-group-component.test.js
@@ -156,15 +156,15 @@ describe('GasPriceButtonGroup Component', function () {
   })
 
   describe('renderButtonContent', () => {
-    it('should render a label if passed a labelKey', () => {
+    it('should render a label if passed a gasEstimateType', () => {
       const renderButtonContentResult = wrapper.instance().renderButtonContent({
-        labelKey: 'mockLabelKey',
+        gasEstimateType: 'SLOW',
       }, {
         className: 'someClass',
       })
       const wrappedRenderButtonContentResult = shallow(renderButtonContentResult)
       assert.equal(wrappedRenderButtonContentResult.childAt(0).children().length, 1)
-      assert.equal(wrappedRenderButtonContentResult.find('.someClass__label').text(), 'mockLabelKey')
+      assert.equal(wrappedRenderButtonContentResult.find('.someClass__label').text(), 'slow')
     })
 
     it('should render a feeInPrimaryCurrency if passed a feeInPrimaryCurrency', () => {
@@ -211,7 +211,7 @@ describe('GasPriceButtonGroup Component', function () {
 
     it('should render all elements if all args passed', () => {
       const renderButtonContentResult = wrapper.instance().renderButtonContent({
-        labelKey: 'mockLabel',
+        gasEstimateType: 'SLOW',
         feeInPrimaryCurrency: 'mockFeeInPrimaryCurrency',
         feeInSecondaryCurrency: 'mockFeeInSecondaryCurrency',
         timeEstimate: 'mockTimeEstimate',