+++ ui/app/components/app/transaction-list-item/transaction-list-item.component.js
@@ -24,7 +24,7 @@ export default class TransactionListItem extends PureComponent {
     showCancelModal: PropTypes.func,
     showCancel: PropTypes.bool,
     hasEnoughCancelGas: PropTypes.bool,
-    showRetry: PropTypes.bool,
+    showSpeedUp: PropTypes.bool,
     isEarliestNonce: PropTypes.bool,
     showFiat: PropTypes.bool,
     token: PropTypes.object,
@@ -177,7 +177,7 @@ export default class TransactionListItem extends PureComponent {
       primaryTransaction,
       showCancel,
       hasEnoughCancelGas,
-      showRetry,
+      showSpeedUp,
       tokenData,
       transactionGroup,
       rpcPrefs,
@@ -233,7 +233,8 @@ export default class TransactionListItem extends PureComponent {
                 <TransactionListItemDetails
                   transactionGroup={transactionGroup}
                   onRetry={this.handleRetry}
-                  showRetry={showRetry}
+                  showSpeedUp={showSpeedUp}
+                  showRetry={getStatusKey(primaryTransaction) === 'failed'}
                   isEarliestNonce={isEarliestNonce}
                   onCancel={this.handleCancel}
                   showCancel={showCancel}