+++ ui/app/pages/send/send.container.js
@@ -62,11 +62,6 @@ import {
   SEND_ROUTE,
 } from '../../helpers/constants/routes'
 
-module.exports = compose(
-  withRouter,
-  connect(mapStateToProps, mapDispatchToProps)
-)(SendEther)
-
 function mapStateToProps (state) {
   return {
     amount: getSendAmount(state),
@@ -140,3 +135,8 @@ function mapDispatchToProps (dispatch) {
     },
   }
 }
+
+export default compose(
+  withRouter,
+  connect(mapStateToProps, mapDispatchToProps)
+)(SendEther)