+++ ui/app/pages/settings/contact-list-tab/edit-contact/edit-contact.component.js
@@ -18,22 +18,29 @@ export default class EditContact extends PureComponent {
     history: PropTypes.object,
     name: PropTypes.string,
     address: PropTypes.string,
+    chainId: PropTypes.string,
     memo: PropTypes.string,
     viewRoute: PropTypes.string,
     listRoute: PropTypes.string,
     setAccountLabel: PropTypes.func,
   }
 
+  static defaultProps = {
+    name: '',
+    address: '',
+    memo: '',
+  }
+
   state = {
-    newName: '',
-    newAddress: '',
-    newMemo: '',
+    newName: this.props.name,
+    newAddress: this.props.address,
+    newMemo: this.props.memo,
     error: '',
   }
 
   render () {
     const { t } = this.context
-    const { history, name, addToAddressBook, removeFromAddressBook, address, memo, viewRoute, listRoute, setAccountLabel } = this.props
+    const { history, name, addToAddressBook, removeFromAddressBook, address, chainId, memo, viewRoute, listRoute, setAccountLabel } = this.props
 
     return (
       <div className="settings-page__content-row address-book__edit-contact">
@@ -43,7 +50,7 @@ export default class EditContact extends PureComponent {
             type="link"
             className="settings-page__address-book-button"
             onClick={() => {
-              removeFromAddressBook(address)
+              removeFromAddressBook(chainId, address)
               history.push(listRoute)
             }}
           >
@@ -59,7 +66,7 @@ export default class EditContact extends PureComponent {
               type="text"
               id="nickname"
               placeholder={this.context.t('addAlias')}
-              value={this.state.newName || name}
+              value={this.state.newName}
               onChange={e => this.setState({ newName: e.target.value })}
               fullWidth
               margin="dense"
@@ -73,8 +80,7 @@ export default class EditContact extends PureComponent {
             <TextField
               type="text"
               id="address"
-              placeholder={address}
-              value={this.state.newAddress || address}
+              value={this.state.newAddress}
               error={this.state.error}
               onChange={e => this.setState({ newAddress: e.target.value })}
               fullWidth
@@ -90,7 +96,7 @@ export default class EditContact extends PureComponent {
               type="text"
               id="memo"
               placeholder={memo}
-              value={this.state.newMemo || memo}
+              value={this.state.newMemo}
               onChange={e => this.setState({ newMemo: e.target.value })}
               fullWidth
               margin="dense"
@@ -109,12 +115,12 @@ export default class EditContact extends PureComponent {
             if (this.state.newAddress !== '' && this.state.newAddress !== address) {
               // if the user makes a valid change to the address field, remove the original address
               if (isValidAddress(this.state.newAddress)) {
-                removeFromAddressBook(address)
+                removeFromAddressBook(chainId, address)
                 addToAddressBook(this.state.newAddress, this.state.newName || name, this.state.newMemo || memo)
                 setAccountLabel(this.state.newAddress, this.state.newName || name)
                 history.push(listRoute)
               } else {
-                this.setState({ error: 'invalid address' })
+                this.setState({ error: this.context.t('invalidAddress') })
               }
             } else {
               // update name